Removing earlier version of blinkiva
