+++ environment.yml
@@ -1,11 +1,17 @@
-name: 2019_icassp_scheibler_blinkyiva
+name: 2019_scheibler_overiva
 dependencies:
   - python=3.7
   - numpy
   - scipy>=0.18.0
+  - mkl-service
   - matplotlib
+  - pandas
+  - ipyparallel
+  - pip
   - pip:
-    - pyroomacoustics=0.1.21
+    - seaborn==0.9.0
+    - gitpython
+    - pyroomacoustics==0.1.23
     - sounddevice
     - samplerate
     - mir_eval
