+++ routines.py
@@ -1,15 +1,15 @@
 # Copyright (c) 2019 Robin Scheibler
-# 
+#
 # Permission is hereby granted, free of charge, to any person obtaining a copy
 # of this software and associated documentation files (the "Software"), to deal
 # in the Software without restriction, including without limitation the rights
 # to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
 # copies of the Software, and to permit persons to whom the Software is
 # furnished to do so, subject to the following conditions:
-# 
+#
 # The above copyright notice and this permission notice shall be included in all
 # copies or substantial portions of the Software.
-# 
+#
 # THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
 # IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
 # FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
@@ -23,66 +23,6 @@ from tkinter import Tk, Button, Label
 
 import sounddevice as sd
 import pyroomacoustics as pra
-
-
-class EncodingBlinkyArray(pra.MicrophoneArray):
-    '''
-    This is an abstraction for the sound to light sensors
-    using frequency domain encoding
-
-    Attributes
-    ----------
-    R: array_like (n_dim, n_sensors)
-        The sensor locations
-    fs: int
-        The sampling frequency
-    freq_lim: tuple length 2
-        Limits of the band in which to filter the data
-    downsampling: int, optional
-        The length of averaging
-    filters: array_like (n_sensors, filter_length)
-        The encoding filters 
-    '''
-
-    def __init__(self, R, fs, freq_lim=None, downsampling=1, filters=None):
-        pra.MicrophoneArray.__init__(self, R, fs)
-        self.filters = filters
-        self.downsampling = downsampling
-        self.nfft = 2 * filters.shape[1]
-        self.filters_fd = np.fft.rfft(filters, axis=-1, n=self.nfft).T / np.sqrt(self.nfft)
-
-    def record(self, signals, fs):
-        '''
-        This simulates the recording of the signals by the microphones
-        and the transformation from sound to light.
-         
-        Parameters
-        ----------
-        signals:
-            An ndarray with as many lines as there are microphones.
-        fs:
-            the sampling frequency of the signals.
-        '''
-
-        self.raw_signals = signals
-
-        X = pra.transform.analysis(
-                signals.T,
-                self.downsampling,
-                self.downsampling,
-                zp_back=self.nfft - self.downsampling,
-                )
-        X /= np.sqrt(self.nfft)
-
-        # frequency domain filtering
-        Y = X * self.filters_fd[None,]
-
-        self.filtered_signals = pra.transform.synthesis(Y, self.downsampling, self.downsampling, zp_back=self.nfft - self.downsampling)
-
-        # power averaging to get the blinky signals
-        self.signals = np.mean(np.abs(Y) ** 2, axis=1).T
-
-        self.power = np.mean(np.abs(X) ** 2, axis=1).T
 
 
 # Now come the GUI part
@@ -105,18 +45,32 @@ class PlaySoundGUI(object):
         self.label.grid(row=nrow, columnspan=2)
         nrow += 1
 
-        self.mix_button = Button(master, text='Mix', command=lambda: self.play(self.mix))
+        self.mix_button = Button(
+            master, text="Mix", command=lambda: self.play(self.mix)
+        )
         self.mix_button.grid(row=nrow, columnspan=2)
         nrow += 1
 
         self.buttons = []
         for i, source in enumerate(self.sources):
-            self.buttons.append(Button(master, text='Source ' + str(i+1), command=lambda src=source: self.play(src)))
+            self.buttons.append(
+                Button(
+                    master,
+                    text="Source " + str(i + 1),
+                    command=lambda src=source: self.play(src),
+                )
+            )
 
             if self.references is not None:
                 self.buttons[-1].grid(row=nrow, column=1)
-                ref_sig = self.references[i,:]
-                self.buttons.append(Button(master, text='Ref ' + str(i+1), command=lambda rs=self.references[i,:]: self.play(rs)))
+                ref_sig = self.references[i, :]
+                self.buttons.append(
+                    Button(
+                        master,
+                        text="Ref " + str(i + 1),
+                        command=lambda rs=self.references[i, :]: self.play(rs),
+                    )
+                )
                 self.buttons[-1].grid(row=nrow, column=0)
 
             else:
@@ -136,49 +90,6 @@ class PlaySoundGUI(object):
         sd.play(0.75 * src / self.sources_max, samplerate=self.fs, blocking=False)
 
 
-
-def generate_filters(n_sensors, length, type='binary'):
-    '''
-    Generates random filters
-
-    Parameters
-    ----------
-    n_sensors: int
-        The number of sensors
-    length: int
-        The length of the filters
-    type: str
-        The type of filter to generate (binary or real)
-
-
-    returns
-    -------
-    filters: array_like (n_sensors, lengths)
-        The filters
-
-    '''
-
-    # the filters have to be symetric to make sure they have a real transfer function
-    if length % 2 == 0:
-        n_half = length // 2
-    else:
-        n_half = length // 2 + 1
-
-    if type == 'binary':
-        F_half = np.random.choice(2, size=(n_sensors, n_half))
-    else:
-        F_half = np.sqrt(np.random.rand(n_sensors, n_half))
-
-    blocks = [F_half]
-    if length % 2 == 0:
-        blocks.append(np.zeros((n_sensors,1)))
-    blocks.append(F_half[:,:0:-1])
-
-    filters = np.concatenate(blocks, axis=1)
-
-    return filters
-
-
 def random_layout(vol_dim, n_mic, offset=None, seed=None):
 
     if seed is not None:
@@ -193,7 +104,7 @@ def random_layout(vol_dim, n_mic, offset=None, seed=None):
     vol_dim = vol_dim + np.array(offset)
 
     points = []
-    for l,u in zip(offset, vol_dim):
+    for l, u in zip(offset, vol_dim):
         points.append(np.random.uniform(l, u, size=n_mic))
 
     if seed is not None:
@@ -203,9 +114,9 @@ def random_layout(vol_dim, n_mic, offset=None, seed=None):
 
 
 def grid_layout(room_dim, n_mic, offset=None, seed=None):
-    '''
+    """
     Place the microphones equispaced on a grid
-    '''
+    """
 
     area = np.prod(room_dim[:2])
     sq_L = np.sqrt(area / n_mic)
@@ -215,9 +126,9 @@ def grid_layout(room_dim, n_mic, offset=None, seed=None):
     hb = n_mic // wb
 
     n = 0
-    x,y = 0, 0
+    x, y = 0, 0
     while n < n_mic:
-        mic_loc.append([x, y, 0.])
+        mic_loc.append([x, y, 0.0])
 
         n = n + 1
         if n % wb == 0:
@@ -229,7 +140,7 @@ def grid_layout(room_dim, n_mic, offset=None, seed=None):
 
     # center the microphones (x-y plan only)
     for i in range(2):
-        mic_loc[i,:] += (room_dim[i] - mic_loc[i,:].max()) / 2
+        mic_loc[i, :] += (room_dim[i] - mic_loc[i, :].max()) / 2
 
     if offset is not None:
         mic_loc += np.array([offset]).T
@@ -244,20 +155,20 @@ def grid_layout(room_dim, n_mic, offset=None, seed=None):
 
 
 def semi_circle_layout(center, angle, distance, n, rot=None, seed=None):
-    '''
+    """
     Places n points on a semi circle covering an angle, at a distance from center
-    '''
+    """
 
     center = np.array(center)
 
     angles = np.linspace(0, angle, n) + rot
 
-    v = np.array([np.cos(angles), np.sin(angles),]) * distance
+    v = np.array([np.cos(angles), np.sin(angles)]) * distance
 
     if center.shape[0] == 3:
-        v = np.concatenate([v, np.zeros((1,n))], axis=0)
+        v = np.concatenate([v, np.zeros((1, n))], axis=0)
 
-    v += center[:,None]
+    v += center[:, None]
 
     if seed is not None:
         rng_state = np.random.get_state()
@@ -282,18 +193,17 @@ def gm_layout(n, centers, std=None, weights=None, seed=None):
     if weights is None:
         weights = np.ones(centers.shape[1]) / centers.shape[1]
 
-    locs = [] 
+    locs = []
 
     rep = math.ceil(n / centers.shape[1])
     c_list = np.repeat(np.arange(centers.shape[1]), rep)[:n]
 
     for c in c_list:
 
-        loc = centers[:,c] + np.random.randn(centers.shape[0]) * std
+        loc = centers[:, c] + np.random.randn(centers.shape[0]) * std
         locs.append(loc)
 
     if seed is not None:
         np.random.set_state(rng_state)
 
     return np.array(locs).T
-
